requirements.txt
numpy
praat-parselmouth
pydub
static-ffmpeg
moviepy